root: Clean 'java.*' from the common source sets
Common source should not directly depend on 'java.*' classes.
Though this is effectively a non-issue because we do have JVM as one of
the targets, it's still an inappropriate thing to do.
